analysis-worker/requirements.txt
anthropic>=0.18.0
pillow>=10.0.0
pytesseract>=0.3.10
click>=8.1.0
